pyproject.toml
[project]
name = "hwaui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "eventlet==0.36.1",
    "fastapi>=0.135.1",
    "flask==3.0.3",
    "flask-socketio==5.3.6",
    "google-genai>=1.66.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "joblib==1.4.0",
    "lightgbm==4.3.0",
    "numpy==1.26.4",
    "opencv-python-headless>=4.11.0.86",
    "paho-mqtt==1.6.1",
    "pandas==2.2.1",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "scikit-learn==1.4.2",
    "seaborn>=0.13.2",
    "supabase>=2.28.0",
    "torch==2.2.0",
    "torchvision>=0.17.0",
    "tqdm>=4.67.3",
    "ultralytics>=8.4.21",
    "uvicorn>=0.41.0",
    "xgboost==2.0.3",
]
